+++ TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/8x_ShaderGraph/8212_Fullscreen/CustomPP1_8210.cs
@@ -1,0 +1,43 @@
+using UnityEngine;
+using UnityEngine.Rendering;
+using UnityEngine.Rendering.HighDefinition;
+using System;
+
+[Serializable, VolumeComponentMenu("Post-processing/Custom/8212_CustomPP1")]
+public sealed class CustomPP1_8210 : CustomPostProcessVolumeComponent, IPostProcessComponent
+{
+    [Tooltip("Controls the intensity of the effect.")]
+    public ClampedFloatParameter intensity = new ClampedFloatParameter(0f, 0f, 1f);
+
+    Material m_Material;
+
+    public bool IsActive() => m_Material != null && intensity.value > 0f;
+
+    // Do not forget to add this post process in the Custom Post Process Orders list (Project Settings > Graphics > HDRP Settings).
+    public override CustomPostProcessInjectionPoint injectionPoint => CustomPostProcessInjectionPoint.BeforePostProcess;
+
+    const string kShaderName = "Hidden/Shader/8212_CustomPP1";
+
+    public override void Setup()
+    {
+        if (Shader.Find(kShaderName) != null)
+            m_Material = new Material(Shader.Find(kShaderName));
+        else
+            Debug.LogError($"Unable to find shader '{kShaderName}'. Post Process Volume 8212_CustomPP1 is unable to load.");
+    }
+
+    public override void Render(CommandBuffer cmd, HDCamera camera, RTHandle source, RTHandle destination)
+    {
+        if (m_Material == null)
+            return;
+
+        m_Material.SetFloat("_Intensity", intensity.value);
+        m_Material.SetTexture("_MainTex", source);
+        HDUtils.DrawFullScreen(cmd, m_Material, destination, shaderPassId: 0);
+    }
+
+    public override void Cleanup()
+    {
+        CoreUtils.Destroy(m_Material);
+    }
+}