+++ com.unity.render-pipelines.universal/Runtime/2D/Passes/UpscalePass.cs
@@ -35,7 +35,7 @@ namespace UnityEngine.Rendering.Universal
 
         public override void Execute(ScriptableRenderContext context, ref RenderingData renderingData)
         {
-            var cmd = CommandBufferPool.Get();
+            var cmd = renderingData.commandBuffer;
             using (new ProfilingScope(cmd, m_ProfilingScope))
             {
                 CoreUtils.SetRenderTarget(cmd, m_UpscaleHandle,
@@ -43,8 +43,6 @@ namespace UnityEngine.Rendering.Universal
                     ClearFlag.None, Color.clear);
                 Blit(cmd, m_Source, m_UpscaleHandle);
             }
-            context.ExecuteCommandBuffer(cmd);
-            CommandBufferPool.Release(cmd);
         }
     }
 }