+++ com.unity.shadergraph/Editor/AssemblyInfo.cs
@@ -10,3 +10,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.RenderPipelines.HighDefinition.Editor")]
 [assembly: InternalsVisibleTo("Unity.VisualEffectGraph.Editor")]
 [assembly: InternalsVisibleTo("Unity.Industrial.Materials.AVRD.Editor")]
+[assembly: InternalsVisibleTo("Unity.VisualEffectGraph.EditorTests")]