+++ com.unity.shadergraph/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package are documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [14.0.2] - 2021-02-04
+
+### Fixed
+ - Fixed ShaderGraph pixel and screen coordinates to work correctly with render scale [1387468]
+
 ## [14.0.1] - 2021-12-07
 
 ### Added